requirements.txt
langchain==1.2.12
langchain-core==1.2.19
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-openai==1.1.11
streamlit==1.49.1
ddgs==9.11.4
arxiv==2.4.1
wikipedia==1.4.0
python-dotenv==1.2.2
ipykernel
pypdf
bs4
pymupdf
langchain-text-splitters
chromadb
sentence-transformers
langchain-huggingface
faiss-cpu
langchain-chroma
duckdb
pandas
openai
duckduckgo-search
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub